Remove redundant test
There's only a confirm step that always sets the "allow EA versions" config option.
The behavior doesn't differ anymore if the user switches to another release-channel or version. The option is always set and the repair flow is the same.
The tests were mostly identical, except for this step, so we can safely remove this test now.